mengganti tipe data float jadi int

+++ Soal1.py
@@ -4,7 +4,7 @@
 #If % is >=70 then Good performance
 #If % is >=60 then average performance
 
-nilai = float(input("Masukkan nilainya: "))
+nilai = int(input("Masukkan nilainya: "))
 
 if nilai >= 90:
     print("Excellent")
